Refactor setDeliveryPilulier method to accept delivery times
The setDeliveryPilulier method in the AidantController has been modified to accept delivery times as parameters. The method now retrieves the delivery times from the request and calls the setDeliveryTime method in the Pilulier entity, passing the time values. If any of the required time parameters are missing, an error response is returned. The setDeliveryTime method in the Pilulier entity has also been updated to include the delivery times in the Python API URL for setting the delivery times. Any errors that occur during the process are properly handled and appropriate messages are displayed.